Don't set list-visible unless the widget is visible
list-visible is a special property that means the
list part of the chooser widget is visible.  We need
to make sure we only report it visible when the
chooser itself is visible otherwise the property is
a lie.

This fixes focus handling when the conversation list
isn't visible to begin with.