manage tasks outside of task list
The task list isn't very good for tracking tasks as
they come and go, since it replies on the groaty details
of toggle buttons and widget activation and such.

We now use it soley for display purposes when necessary
and otherwise, keep track of the tasks in a separate list.

Also, differentiate single authentication modes from
multiple authentication modes, so we don't show the task list
when we don't need to.