gdm: always run gdm on VT1
This drops automatic VT allocation schemes for the initial display in
favour of a compile time hardcoded VT assignment. The automatic
allocation schemes are inherently racy since a simpler output to the
console might already influence it and result in gdm running on another
than the intended VT.

This patch adds a --with-initial-vt= switch to configure which may be
used to set the VT gdm will run the initial server on. It defaults to 1.

https://fedoraproject.org/wiki/Features/DisplayManagerRework

https://bugzilla.gnome.org/show_bug.cgi?id=511168