slave: run PreSession before starting the session
We currently run the PreSession script after we've already started
loading the session. This means zenity dialogs and the like won't
have time to get user confirmation before login continues.

This commit calls the PreSession script a bit earlier, after the
session is opened, but before it's started.  Following this commit,
admins will be delay starting the session until a program of their
choosing completes, or even cancel the log in process all together.