xdcmp: fix selecting remote hosts from chooser
The XDMCP indirect query is supposed to be forwarded by
gdm_xdmcp_send_forward_query() from daemon/gdm-xdmcp-display-factory.c:

XdmcpFlush (factory->priv->socket_fd,
            &factory->priv->buf,
            (XdmcpNetaddr)gdm_address_peek_sockaddr_storage
            (ic->chosen_address),
            (int)gdm_sockaddr_len
            (gdm_address_peek_sockaddr_storage
             (ic->chosen_address)));

However, checking the return value of XdmcpFlush() we can see
the call fails with "Invalid Argument". The address (ic->chosen_address)
selected is correct but the port is 0 instead of 177, so the indirect
query is never forwarded to the selected host.

The reason for this is because the chosen_address is set in
on_hostname_selected() from daemon/gdm-xdmcp-display-factory.c via
getaddrinfo() but the service field passed to getaddrinfo() is NULL,
while it should be the actual XDMCP port ("177").

This commit ensures the correct port is passed to getaddrinfo and
fixes the chooser when choosing remote machines.