manager: collect more details about bus sender
In the future we're going to need to know more details about the
sender to know how to move forward (such as seat id, session id,
if it's remote, etc) in order to create a transient session soley
for reauthentication.

To prepare for that future, this commit adds the necessary
functionality to get_display_and_details_for_bus_sender.