login-window: only focus window if there's a valid user timestamp
set_focus() is called any time the login window is reset, including at
startup.  set_focus() re-presents the window in the case it lost focus.
At startup it won't have lost focus, and gdk_window_focus will cause
window manager spew because the user time is 0 (since we don't start
the greeter using start up notification, and the user hasn't had a
chance to interact with the display yet).

This commit explicitly checks for user time of 0 and avoids presenting
the window in that case.