xdmcp-greeter-display: gut into the useless stub it is
This commit drops the finish implementation from the
xdmcp-greeter-display subclass.  It's mostly just chaining up, but it
also calls "unmanage" which is strange and probably wrong.

At the same time, it turns out pretty much all the other code in the
file is superfluous, too, so this commit kills it all dead.

A future commit will probably get rid of the file entirely.