util: add wrapper for launching X session
This commit adds a utility for launching an X server and session
together.

This simple utility works a lot like xinit, but makes use of
modern X features (like -displayfd) and has GDM specific integration
(like reading from gdm configuration, and optionally running through
 /etc/gdm/Xsession)

The eventual idea is to get the main GDM code out of the X server launching
business and instead farm the logic off to an unprivileged helper
that gets run within the logind session of the user.