session: jump to VT up front for X sessions
It turns out X doesn't work to well when run without being
setuid, right now.  This is because it relies on being the
foreground to leverage logind for devices.

As a workaround, this commit jumps to the VT up front before
starting X.  This should be flicker free now, because we put
it in KD_GRAPHICS mode first.