manager: don't muck with the session type and class of stub display (autologin)
Normally at startup a display is created of class "greeter" for the
login screen to use.  That display isn't used in the case automatic
login is enabled.  There's some iffy code to try to change the class
from "greeter" to "user" in the automatic login case.  That code is
wrong, because in the automatic login case a new display is created
specifically for the user session.  The greeter session is abandoned,
instead of used, so mucking with it's class and type is wrong.

This commit removes the code that does that mucking.