gdm-session: update session type when the session defaults change
If a user has never logged in before and they login using autologin,
then we don't select the correct session type if they end up
with a wayland session.  This is because we fail to initialize
the session type based on the default session type.

This commit changes the code to update the session type anytime the
session defaults are reset.

https://bugzilla.gnome.org/show_bug.cgi?id=761785