manager: update session-id property when reusing display
If a display starts out its life as a greeter display, and then
gets reused for the user session, we need to update the session-id
property on the display to match its new session.

This is important so the reauthentication process match the
session with existing display and run in the proper context.