legacy-display: ensure X server is dead when respawning
If the user logs out we may detect the session dying before
we detect the X server dying.  If that happens, then we
currently don't wait for the X server to exit before starting
a new X server.

This commit makes sure to explicitly kill the X server and wait
for its death before proceeding.

https://bugzilla.gnome.org/show_bug.cgi?id=788786