manager: don't record wtmp entries when user unknown
Right now we end up writing wtmp entries for the login screen
user into /var/log/wtmp, but with a funky username of "(unknown)".

The login screen session shouldn't get a wtmp entry, and we
shouldn't ever add wtmp entries for sessions we don't know the username
for.

This commit fixes that.

https://bugzilla.gnome.org/show_bug.cgi?id=788784