daemon: handle upgrades from RHEL 7
RHEL 7 users need to stay on X if they were using X,
and they need to stay on gnome-classic if they were using
gnome-classic.

This commit examines the user's config to deduce whether
or not they were using RHEL 7 and in the event they were
try to get the right settings.